se modifica la request update, se generan cambios en la documentacion

+++ database/migrations/2021_12_16_164844_create_settings_table.php
@@ -16,7 +16,7 @@ class CreateSettingsTable extends Migration
         Schema::create('settings', function (Blueprint $table) {
             $table->id();
             $table->string('name')->unique();
-            $table->integer('value');
+            $table->double('value');
             $table->timestamps();
 
         });
